+++ src/refactoring/hooks/useProductManager.ts
@@ -1,4 +1,5 @@
 import { useState } from "react";
+
 import { Discount, Product } from "../../types";
 
 interface Props {
@@ -11,6 +12,7 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
   const [editingProduct, setEditingProduct] = useState<Product | null>(null);
   const [newDiscount, setNewDiscount] = useState<Discount>(initialDiscount);
 
+  // editing product의 내용을 업데이트하는 함수
   const handleProductUpdate = <K extends keyof Product>(
     productId: String,
     key: K,
@@ -22,24 +24,29 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
     }
   };
 
+  // editing product를 업데이트하는 함수
   const handleEditProduct = (product: Product) => {
     setEditingProduct({ ...product });
   };
 
-  const handleProductNameUpdate = (productId: string, newName: string) => {
+  // editing product의 name을 업데이트하는 함수
+  const handleNameUpdate = (productId: string, newName: string) => {
     handleProductUpdate(productId, "name", newName);
   };
 
+  // editing product의 price를 업데이트하는 함수
   const handlePriceUpdate = (productId: string, newPrice: number) => {
     if (newPrice < 0) return console.error("price cannot be navigate");
     handleProductUpdate(productId, "price", newPrice);
   };
 
+  // editing product의 stock을 업데이트하는 함수
   const handleStockUpdate = (productId: string, newStock: number) => {
     if (newStock < 0) return console.error("stock cannot be navigate");
     handleProductUpdate(productId, "stock", newStock);
   };
 
+  // 수정한 프로덕트를 실제 product에 반영하는 함수
   const updateProduct = (product: Product) => {
     onProductUpdate(product);
     if (product.id === editingProduct?.id) {
@@ -47,6 +54,7 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
     }
   };
 
+  // 압력중인 새로운 할인을 editing product에 추가하는 함수 (실제 product에 바로 반영됨))
   const handleAddDiscount = () => {
     if (editingProduct) {
       const updatedProduct = {
@@ -58,6 +66,7 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
     }
   };
 
+  // editing product에서 특정 할인을 제거하는 함수 (실제 product에 바로 반영됨)
   const handleRemoveDiscount = (index: number) => {
     if (editingProduct) {
       const updatedProduct = {
@@ -68,6 +77,7 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
     }
   };
 
+  // 수정중인 product정보 저장
   const handleEditComplete = () => {
     if (editingProduct) {
       onProductUpdate(editingProduct);
@@ -75,12 +85,14 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
     }
   };
 
+  // 할인의 개수제한(quantity)를 업데이트하는 함수
   const handleDiscountQuantityUpdate = (newQuantity: number) => {
     if (newQuantity > 0) {
       setNewDiscount({ ...newDiscount, quantity: newQuantity });
     }
   };
 
+  // 할인의 할인율(rate)를 업데이트하는 함수
   const handleDiscountRateUpdate = (newRate: number) => {
     setNewDiscount({ ...newDiscount, rate: newRate / 100 });
   };
@@ -88,7 +100,7 @@ export const useProductManager = ({ onProductUpdate }: Props) => {
   return {
     editingProduct,
     handleEditProduct,
-    handleProductNameUpdate,
+    handleNameUpdate,
     handlePriceUpdate,
     handleStockUpdate,
     handleAddDiscount,
